+++ src/database/operations/changemakerProposals/createChangemakerProposal.ts
@@ -1,4 +1,5 @@
 import { generateCreateOrUpdateItemOperation } from '../generators';
+import { decorateWithFileDownloadUrls } from '../../../decorators/changemakerProposal';
 import type {
 	ChangemakerProposal,
 	WritableChangemakerProposal,
@@ -8,6 +9,11 @@ const createChangemakerProposal = generateCreateOrUpdateItemOperation<
 	ChangemakerProposal,
 	WritableChangemakerProposal,
 	[]
->('changemakersProposals.insertOne', ['changemakerId', 'proposalId'], []);
+>(
+	'changemakersProposals.insertOne',
+	['changemakerId', 'proposalId'],
+	[],
+	decorateWithFileDownloadUrls,
+);
 
 export { createChangemakerProposal };